Add credential search.
We allow filtering via a regular expression on the target-name, since that has both the user and the service in it, and saves the client the trouble of downcasting the credential.

We deduce whether each credential found could be specified, and if so we make it a specifier.

Credentials created in tests now have their service names start with `test-`.

Also provide an additional example which demonstrates search by using it to clean up abandoned test entries.